HSEARCH-4303 Extract an abstract base for AutomaticIndexingSingleAssociationIT
So that we can create more tests for other types of single-valued
associations (@ManyToOne, ...).
