--- integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/AutomaticIndexingSingleAssociationIT.java
@@ -1,1255 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Stream;
-import javax.persistence.Basic;
-import javax.persistence.ElementCollection;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-import javax.persistence.Transient;
-
-import org.hibernate.search.mapper.pojo.automaticindexing.ReindexOnUpdate;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexingDependency;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.ObjectPath;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.PropertyValue;
-import org.hibernate.search.util.impl.integrationtest.mapper.orm.OrmUtils;
-import org.hibernate.search.util.impl.test.annotation.TestForIssue;
-
-import org.junit.Test;
-
-/**
- * Test automatic indexing based on Hibernate ORM entity events
- * when a ToOne association is involved.
- * <p>
- * See {@link AbstractAutomaticIndexingAssociationIT} for more details on how this test is designed.
- */
-public class AutomaticIndexingSingleAssociationIT extends AbstractAutomaticIndexingAssociationIT<
-		AutomaticIndexingSingleAssociationIT.IndexedEntity,
-		AutomaticIndexingSingleAssociationIT.ContainingEntity,
-		AutomaticIndexingSingleAssociationIT.ContainedEntity
-		> {
-
-	public AutomaticIndexingSingleAssociationIT() {
-		super( new ModelPrimitives() );
-	}
-
-	@Test
-	public void directAssociationUpdate_indexedEmbedded() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> { } );
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 2 );
-			containedEntity.setIndexedField( "initialValue" );
-
-			entity1.setContainedIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsIndexedEmbedded( entity1 );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "containedIndexedEmbedded", b2 -> b2
-									.field( "indexedField", "initialValue" )
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 3 );
-			containedEntity.setIndexedField( "updatedValue" );
-
-			entity1.getContainedIndexedEmbedded().setContainingAsIndexedEmbedded( null );
-			entity1.setContainedIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsIndexedEmbedded( entity1 );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "containedIndexedEmbedded", b2 -> b2
-									.field( "indexedField", "updatedValue" )
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			entity1.getContainedIndexedEmbedded().setContainingAsIndexedEmbedded( null );
-			entity1.setContainedIndexedEmbedded( null );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> { } );
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Test that updating a non-IndexedEmbedded association in an entity
-	 * whose other properties are indexed
-	 * does not trigger reindexing of the entity.
-	 */
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-3199")
-	public void directAssociationUpdate_nonIndexedEmbedded() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> { } );
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 2 );
-			containedEntity.setNonIndexedField( "initialValue" );
-
-			entity1.setContainedNonIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsNonIndexedEmbedded( entity1 );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 3 );
-			containedEntity.setNonIndexedField( "updatedValue" );
-
-			entity1.getContainedNonIndexedEmbedded().setContainingAsNonIndexedEmbedded( null );
-			entity1.setContainedNonIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsNonIndexedEmbedded( entity1 );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			entity1.getContainedNonIndexedEmbedded().setContainingAsNonIndexedEmbedded( null );
-			entity1.setContainedNonIndexedEmbedded( null );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Test that updating an IndexedEmbedded association in an indexed entity
-	 * does trigger reindexing of the entity
-	 * if the association is marked with ReindexOnUpdate = SHALLOW.
-	 */
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-4001")
-	public void directAssociationUpdate_indexedEmbeddedShallowReindexOnUpdate() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> { } );
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 2 );
-			containedEntity.setIndexedField( "initialValue" );
-
-			entity1.setContainedIndexedEmbeddedShallowReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "containedIndexedEmbeddedShallowReindexOnUpdate", b2 -> b2
-									.field( "indexedField", "initialValue" )
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 3 );
-			containedEntity.setIndexedField( "updatedValue" );
-
-			entity1.setContainedIndexedEmbeddedShallowReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "containedIndexedEmbeddedShallowReindexOnUpdate", b2 -> b2
-									.field( "indexedField", "updatedValue" )
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			entity1.setContainedIndexedEmbeddedShallowReindexOnUpdate( null );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> { } );
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Test that updating an IndexedEmbedded association in an indexed entity
-	 * does not trigger reindexing of the entity
-	 * if the association is marked with ReindexOnUpdate = NO.
-	 */
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-3206")
-	public void directAssociationUpdate_indexedEmbeddedNoReindexOnUpdate() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> { } );
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 2 );
-			containedEntity.setIndexedField( "initialValue" );
-
-			entity1.setContainedIndexedEmbeddedNoReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 3 );
-			containedEntity.setIndexedField( "updatedValue" );
-
-			entity1.setContainedIndexedEmbeddedNoReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.get( IndexedEntity.class, 1 );
-
-			entity1.setContainedIndexedEmbeddedNoReindexOnUpdate( null );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	@Test
-	public void indirectAssociationUpdate_indexedEmbedded() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			ContainingEntity containingEntity1 = new ContainingEntity();
-			containingEntity1.setId( 2 );
-			entity1.setChild( containingEntity1 );
-			containingEntity1.setParent( entity1 );
-
-			ContainingEntity deeplyNestedContainingEntity = new ContainingEntity();
-			deeplyNestedContainingEntity.setId( 3 );
-			containingEntity1.setChild( deeplyNestedContainingEntity );
-			deeplyNestedContainingEntity.setParent( containingEntity1 );
-
-			session.persist( deeplyNestedContainingEntity );
-			session.persist( containingEntity1 );
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 4 );
-			containedEntity.setIndexedField( "initialValue" );
-
-			containingEntity1.setContainedIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsIndexedEmbedded( containingEntity1 );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.objectField( "containedIndexedEmbedded", b3 -> b3
-											.field( "indexedField", "initialValue" )
-									)
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 5 );
-			containedEntity.setIndexedField( "updatedValue" );
-
-			containingEntity1.getContainedIndexedEmbedded().setContainingAsIndexedEmbedded( null );
-			containingEntity1.setContainedIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsIndexedEmbedded( containingEntity1 );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.objectField( "containedIndexedEmbedded", b3 -> b3
-											.field( "indexedField", "updatedValue" )
-									)
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value that is too deeply nested to matter (it's out of the IndexedEmbedded scope)
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity deeplyNestedContainingEntity1 = session.get( ContainingEntity.class, 3 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 6 );
-			containedEntity.setIndexedField( "outOfScopeValue" );
-
-			deeplyNestedContainingEntity1.setContainedIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsIndexedEmbedded( deeplyNestedContainingEntity1 );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			containingEntity1.getContainedIndexedEmbedded().setContainingAsIndexedEmbedded( null );
-			containingEntity1.setContainedIndexedEmbedded( null );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Test that updating a non-IndexedEmbedded association in an entity
-	 * whose properties are otherwise used in an IndexedEmbedded from an indexed entity
-	 * does not trigger reindexing of the indexed entity.
-	 */
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-3199")
-	public void indirectAssociationUpdate_nonIndexedEmbedded() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			ContainingEntity containingEntity1 = new ContainingEntity();
-			containingEntity1.setId( 2 );
-			entity1.setChild( containingEntity1 );
-			containingEntity1.setParent( entity1 );
-
-			session.persist( containingEntity1 );
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 4 );
-			containedEntity.setIndexedField( "initialValue" );
-
-			containingEntity1.setContainedNonIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsNonIndexedEmbedded( containingEntity1 );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 5 );
-			containedEntity.setIndexedField( "updatedValue" );
-
-			containingEntity1.getContainedNonIndexedEmbedded().setContainingAsNonIndexedEmbedded( null );
-			containingEntity1.setContainedNonIndexedEmbedded( containedEntity );
-			containedEntity.setContainingAsNonIndexedEmbedded( containingEntity1 );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			containingEntity1.getContainedNonIndexedEmbedded().setContainingAsNonIndexedEmbedded( null );
-			containingEntity1.setContainedNonIndexedEmbedded( null );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Test that updating an IndexedEmbedded association in an entity
-	 * that is IndexedEmbedded in an indexed entity
-	 * does trigger reindexing of the indexed entity
-	 * if the first association is marked with ReindexOnUpdate = SHALLOW.
-	 */
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-4001")
-	public void indirectAssociationUpdate_indexedEmbeddedShallowReindexOnUpdate() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			ContainingEntity containingEntity1 = new ContainingEntity();
-			containingEntity1.setId( 2 );
-			entity1.setChild( containingEntity1 );
-			containingEntity1.setParent( entity1 );
-
-			session.persist( containingEntity1 );
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 4 );
-			containedEntity.setIndexedField( "initialValue" );
-
-			containingEntity1.setContainedIndexedEmbeddedShallowReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.objectField( "containedIndexedEmbeddedShallowReindexOnUpdate", b3 -> b3
-											.field( "indexedField", "initialValue" )
-									)
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 5 );
-			containedEntity.setIndexedField( "updatedValue" );
-
-			containingEntity1.setContainedIndexedEmbeddedShallowReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.objectField( "containedIndexedEmbeddedShallowReindexOnUpdate", b3 -> b3
-											.field( "indexedField", "updatedValue" )
-									)
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			containingEntity1.setContainedIndexedEmbeddedShallowReindexOnUpdate( null );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Test that updating an IndexedEmbedded association in an entity
-	 * that is IndexedEmbedded in an indexed entity
-	 * does not trigger reindexing of the indexed entity
-	 * if the first association is marked with ReindexOnUpdate = NO.
-	 */
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-3206")
-	public void indirectAssociationUpdate_indexedEmbeddedNoReindexOnUpdate() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			ContainingEntity containingEntity1 = new ContainingEntity();
-			containingEntity1.setId( 2 );
-			entity1.setChild( containingEntity1 );
-			containingEntity1.setParent( entity1 );
-
-			session.persist( containingEntity1 );
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 4 );
-			containedEntity.setIndexedField( "initialValue" );
-
-			containingEntity1.setContainedIndexedEmbeddedNoReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 5 );
-			containedEntity.setIndexedField( "updatedValue" );
-
-			containingEntity1.setContainedIndexedEmbeddedNoReindexOnUpdate( containedEntity );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			containingEntity1.setContainedIndexedEmbeddedNoReindexOnUpdate( null );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Test that updating an association in an entity
-	 * when this association is a component of a path in a @IndexingDependency.derivedFrom attribute
-	 * does trigger reindexing of the indexed entity.
-	 */
-	@Test
-	public void indirectAssociationUpdate_usedInCrossEntityDerivedProperty() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-
-			ContainingEntity containingEntity1 = new ContainingEntity();
-			containingEntity1.setId( 2 );
-			entity1.setChild( containingEntity1 );
-			containingEntity1.setParent( entity1 );
-
-			ContainingEntity deeplyNestedContainingEntity = new ContainingEntity();
-			deeplyNestedContainingEntity.setId( 3 );
-			containingEntity1.setChild( deeplyNestedContainingEntity );
-			deeplyNestedContainingEntity.setParent( containingEntity1 );
-
-			session.persist( deeplyNestedContainingEntity );
-			session.persist( containingEntity1 );
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 4 );
-			containedEntity.setFieldUsedInCrossEntityDerivedField1( "field1_initialValue" );
-			containedEntity.setFieldUsedInCrossEntityDerivedField2( "field2_initialValue" );
-
-			containingEntity1.setContainedUsedInCrossEntityDerivedProperty( containedEntity );
-			containedEntity.setContainingAsUsedInCrossEntityDerivedProperty( containingEntity1 );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.field(
-											"crossEntityDerivedField",
-											"field1_initialValue field2_initialValue"
-									)
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 5 );
-			containedEntity.setFieldUsedInCrossEntityDerivedField1( "field1_updatedValue" );
-			containedEntity.setFieldUsedInCrossEntityDerivedField2( "field2_updatedValue" );
-
-			containingEntity1.getContainedUsedInCrossEntityDerivedProperty().setContainingAsUsedInCrossEntityDerivedProperty( null );
-			containingEntity1.setContainedUsedInCrossEntityDerivedProperty( containedEntity );
-			containedEntity.setContainingAsUsedInCrossEntityDerivedProperty( containingEntity1 );
-
-			session.persist( containedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.field(
-											"crossEntityDerivedField",
-											"field1_updatedValue field2_updatedValue"
-									)
-							)
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test adding a value that is too deeply nested to matter (it's out of the path)
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity deeplyNestedContainingEntity1 = session.get( ContainingEntity.class, 3 );
-
-			ContainedEntity containedEntity = new ContainedEntity();
-			containedEntity.setId( 6 );
-			containedEntity.setFieldUsedInCrossEntityDerivedField1( "field1_outOfScopeValue" );
-			containedEntity.setFieldUsedInCrossEntityDerivedField2( "field2_outOfScopeValue" );
-
-			deeplyNestedContainingEntity1.setContainedUsedInCrossEntityDerivedProperty( containedEntity );
-			containedEntity.setContainingAsUsedInCrossEntityDerivedProperty( deeplyNestedContainingEntity1 );
-
-			session.persist( containedEntity );
-
-			// Do not expect any work
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test removing a value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainingEntity containingEntity1 = session.get( ContainingEntity.class, 2 );
-
-			containingEntity1.getContainedUsedInCrossEntityDerivedProperty().setContainingAsUsedInCrossEntityDerivedProperty( null );
-			containingEntity1.setContainedUsedInCrossEntityDerivedProperty( null );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "1", b -> b
-							.objectField( "child", b2 -> { } )
-					);
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	private static class ModelPrimitives
-			implements AssociationModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity> {
-
-		@Override
-		public String getIndexName() {
-			return IndexedEntity.INDEX;
-		}
-
-		@Override
-		public boolean isMultiValuedAssociation() {
-			return false;
-		}
-
-		@Override
-		public Class<IndexedEntity> getIndexedClass() {
-			return IndexedEntity.class;
-		}
-
-		@Override
-		public Class<ContainingEntity> getContainingClass() {
-			return ContainingEntity.class;
-		}
-
-		@Override
-		public Class<ContainedEntity> getContainedClass() {
-			return ContainedEntity.class;
-		}
-
-		@Override
-		public IndexedEntity newIndexed(int id) {
-			IndexedEntity entity = new IndexedEntity();
-			entity.setId( id );
-			return entity;
-		}
-
-		@Override
-		public ContainingEntity newContaining(int id) {
-			ContainingEntity entity = new ContainingEntity();
-			entity.setId( id );
-			return entity;
-		}
-
-		@Override
-		public ContainedEntity newContained(int id) {
-			ContainedEntity entity = new ContainedEntity();
-			entity.setId( id );
-			return entity;
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainingEntity, String> containingEntityNonIndexedField() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setNonIndexedField );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainingEntity, ContainingEntity> child() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setChild );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainingEntity, ContainingEntity> parent() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setParent );
-		}
-
-		@Override
-		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbedded() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbedded );
-		}
-
-		@Override
-		public PropertyAccessor<ContainedEntity, ContainingEntity> containingAsIndexedEmbedded() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setContainingAsIndexedEmbedded );
-		}
-
-		@Override
-		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbeddedShallowReindexOnUpdate );
-		}
-
-		@Override
-		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbeddedNoReindexOnUpdate );
-		}
-
-		@Override
-		public PropertyAccessor<ContainingEntity, ContainedEntity> containedUsedInCrossEntityDerivedProperty() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedUsedInCrossEntityDerivedProperty );
-		}
-
-		@Override
-		public PropertyAccessor<ContainedEntity, ContainingEntity> containingAsUsedInCrossEntityDerivedProperty() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setContainingAsUsedInCrossEntityDerivedProperty );
-		}
-
-		@Override
-		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbeddedWithCast() {
-			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbeddedWithCast );
-		}
-
-		@Override
-		public PropertyAccessor<ContainedEntity, ContainingEntity> containingAsIndexedEmbeddedWithCast() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setContainingAsIndexedEmbeddedWithCast );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainedEntity, String> indexedField() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setIndexedField );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainedEntity, String> nonIndexedField() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setNonIndexedField );
-		}
-
-		@Override
-		public MultiValuedPropertyAccessor<ContainedEntity, String, List<String>> indexedElementCollectionField() {
-			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
-					ContainedEntity::getIndexedElementCollectionField,
-					ContainedEntity::setIndexedElementCollectionField );
-		}
-
-		@Override
-		public MultiValuedPropertyAccessor<ContainedEntity, String, List<String>> nonIndexedElementCollectionField() {
-			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
-					ContainedEntity::getNonIndexedElementCollectionField,
-					ContainedEntity::setNonIndexedElementCollectionField );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInContainedDerivedField1() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInContainedDerivedField1 );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInContainedDerivedField2() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInContainedDerivedField2 );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInCrossEntityDerivedField1() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInCrossEntityDerivedField1 );
-		}
-
-		@Override
-		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInCrossEntityDerivedField2() {
-			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInCrossEntityDerivedField2 );
-		}
-	}
-
-	@Entity(name = "containing")
-	public static class ContainingEntity {
-
-		@Id
-		private Integer id;
-
-		private String nonIndexedField;
-
-		@OneToOne
-		private ContainingEntity parent;
-
-		@OneToOne(mappedBy = "parent")
-		@IndexedEmbedded(includePaths = {
-				"containedIndexedEmbedded.indexedField",
-				"containedIndexedEmbedded.indexedElementCollectionField",
-				"containedIndexedEmbedded.containedDerivedField",
-				"containedIndexedEmbeddedShallowReindexOnUpdate.indexedField",
-				"containedIndexedEmbeddedShallowReindexOnUpdate.indexedElementCollectionField",
-				"containedIndexedEmbeddedShallowReindexOnUpdate.containedDerivedField",
-				"containedIndexedEmbeddedNoReindexOnUpdate.indexedField",
-				"containedIndexedEmbeddedNoReindexOnUpdate.indexedElementCollectionField",
-				"containedIndexedEmbeddedNoReindexOnUpdate.containedDerivedField",
-				"containedIndexedEmbeddedWithCast.indexedField",
-				"crossEntityDerivedField"
-		})
-		private ContainingEntity child;
-
-		@OneToOne
-		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
-		private ContainedEntity containedIndexedEmbedded;
-
-		@OneToOne
-		private ContainedEntity containedNonIndexedEmbedded;
-
-		@OneToOne
-		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
-		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
-		private ContainedEntity containedIndexedEmbeddedShallowReindexOnUpdate;
-
-		@OneToOne
-		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
-		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.NO)
-		private ContainedEntity containedIndexedEmbeddedNoReindexOnUpdate;
-
-		@OneToOne
-		private ContainedEntity containedUsedInCrossEntityDerivedProperty;
-
-		@OneToOne(targetEntity = ContainedEntity.class)
-		@IndexedEmbedded(includePaths = { "indexedField" }, targetType = ContainedEntity.class)
-		private Object containedIndexedEmbeddedWithCast;
-
-		public Integer getId() {
-			return id;
-		}
-
-		public void setId(Integer id) {
-			this.id = id;
-		}
-
-		public String getNonIndexedField() {
-			return nonIndexedField;
-		}
-
-		public void setNonIndexedField(String nonIndexedField) {
-			this.nonIndexedField = nonIndexedField;
-		}
-
-		public ContainingEntity getParent() {
-			return parent;
-		}
-
-		public void setParent(ContainingEntity parent) {
-			this.parent = parent;
-		}
-
-		public ContainingEntity getChild() {
-			return child;
-		}
-
-		public void setChild(ContainingEntity child) {
-			this.child = child;
-		}
-
-		public ContainedEntity getContainedIndexedEmbedded() {
-			return containedIndexedEmbedded;
-		}
-
-		public void setContainedIndexedEmbedded(ContainedEntity containedIndexedEmbedded) {
-			this.containedIndexedEmbedded = containedIndexedEmbedded;
-		}
-
-		public ContainedEntity getContainedNonIndexedEmbedded() {
-			return containedNonIndexedEmbedded;
-		}
-
-		public void setContainedNonIndexedEmbedded(ContainedEntity containedNonIndexedEmbedded) {
-			this.containedNonIndexedEmbedded = containedNonIndexedEmbedded;
-		}
-
-		public ContainedEntity getContainedIndexedEmbeddedShallowReindexOnUpdate() {
-			return containedIndexedEmbeddedShallowReindexOnUpdate;
-		}
-
-		public void setContainedIndexedEmbeddedShallowReindexOnUpdate(
-				ContainedEntity containedIndexedEmbeddedShallowReindexOnUpdate) {
-			this.containedIndexedEmbeddedShallowReindexOnUpdate = containedIndexedEmbeddedShallowReindexOnUpdate;
-		}
-
-		public ContainedEntity getContainedIndexedEmbeddedNoReindexOnUpdate() {
-			return containedIndexedEmbeddedNoReindexOnUpdate;
-		}
-
-		public void setContainedIndexedEmbeddedNoReindexOnUpdate(
-				ContainedEntity containedIndexedEmbeddedNoReindexOnUpdate) {
-			this.containedIndexedEmbeddedNoReindexOnUpdate = containedIndexedEmbeddedNoReindexOnUpdate;
-		}
-
-		public ContainedEntity getContainedUsedInCrossEntityDerivedProperty() {
-			return containedUsedInCrossEntityDerivedProperty;
-		}
-
-		public void setContainedUsedInCrossEntityDerivedProperty(
-				ContainedEntity containedUsedInCrossEntityDerivedProperty) {
-			this.containedUsedInCrossEntityDerivedProperty = containedUsedInCrossEntityDerivedProperty;
-		}
-
-		public Object getContainedIndexedEmbeddedWithCast() {
-			return containedIndexedEmbeddedWithCast;
-		}
-
-		public void setContainedIndexedEmbeddedWithCast(Object containedIndexedEmbeddedWithCast) {
-			this.containedIndexedEmbeddedWithCast = containedIndexedEmbeddedWithCast;
-		}
-
-		@Transient
-		@GenericField
-		@IndexingDependency(derivedFrom = {
-				@ObjectPath({
-						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
-						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField1")
-				}),
-				@ObjectPath({
-						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
-						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField2")
-				})
-		})
-		public Optional<String> getCrossEntityDerivedField() {
-			return containedUsedInCrossEntityDerivedProperty == null
-					? Optional.empty()
-					: computeDerived( Stream.of(
-							containedUsedInCrossEntityDerivedProperty.getFieldUsedInCrossEntityDerivedField1(),
-							containedUsedInCrossEntityDerivedProperty.getFieldUsedInCrossEntityDerivedField2()
-					) );
-		}
-	}
-
-	@Entity(name = "indexed")
-	@Indexed(index = IndexedEntity.INDEX)
-	public static class IndexedEntity extends ContainingEntity {
-
-		static final String INDEX = "IndexedEntity";
-
-	}
-
-	@Entity(name = "contained")
-	public static class ContainedEntity {
-		@Id
-		private Integer id;
-
-		@OneToOne(mappedBy = "containedIndexedEmbedded")
-		private ContainingEntity containingAsIndexedEmbedded;
-
-		@OneToOne(mappedBy = "containedNonIndexedEmbedded")
-		private ContainingEntity containingAsNonIndexedEmbedded;
-
-		@OneToOne(mappedBy = "containedUsedInCrossEntityDerivedProperty")
-		private ContainingEntity containingAsUsedInCrossEntityDerivedProperty;
-
-		@OneToOne(mappedBy = "containedIndexedEmbeddedWithCast", targetEntity = ContainingEntity.class)
-		private Object containingAsIndexedEmbeddedWithCast;
-
-		@Basic
-		@GenericField
-		private String indexedField;
-
-		@ElementCollection
-		@GenericField
-		private List<String> indexedElementCollectionField = new ArrayList<>();
-
-		@Basic
-		@GenericField
-		// Keep this annotation, it should be ignored because the field is not included in the @IndexedEmbedded
-		private String nonIndexedField;
-
-		@ElementCollection
-		@GenericField
-		// Keep this annotation, it should be ignored because the field is not included in the @IndexedEmbedded
-		private List<String> nonIndexedElementCollectionField = new ArrayList<>();
-
-		@Basic // Do not annotate with @GenericField, this would make the test pointless
-		private String fieldUsedInContainedDerivedField1;
-
-		@Basic // Do not annotate with @GenericField, this would make the test pointless
-		private String fieldUsedInContainedDerivedField2;
-
-		@Basic // Do not annotate with @GenericField, this would make the test pointless
-		private String fieldUsedInCrossEntityDerivedField1;
-
-		@Basic // Do not annotate with @GenericField, this would make the test pointless
-		private String fieldUsedInCrossEntityDerivedField2;
-
-		public Integer getId() {
-			return id;
-		}
-
-		public void setId(Integer id) {
-			this.id = id;
-		}
-
-		public ContainingEntity getContainingAsIndexedEmbedded() {
-			return containingAsIndexedEmbedded;
-		}
-
-		public void setContainingAsIndexedEmbedded(ContainingEntity containingAsIndexedEmbedded) {
-			this.containingAsIndexedEmbedded = containingAsIndexedEmbedded;
-		}
-
-		public ContainingEntity getContainingAsNonIndexedEmbedded() {
-			return containingAsNonIndexedEmbedded;
-		}
-
-		public void setContainingAsNonIndexedEmbedded(ContainingEntity containingAsNonIndexedEmbedded) {
-			this.containingAsNonIndexedEmbedded = containingAsNonIndexedEmbedded;
-		}
-
-		public ContainingEntity getContainingAsUsedInCrossEntityDerivedProperty() {
-			return containingAsUsedInCrossEntityDerivedProperty;
-		}
-
-		public void setContainingAsUsedInCrossEntityDerivedProperty(
-				ContainingEntity containingAsUsedInCrossEntityDerivedProperty) {
-			this.containingAsUsedInCrossEntityDerivedProperty = containingAsUsedInCrossEntityDerivedProperty;
-		}
-
-		public Object getContainingAsIndexedEmbeddedWithCast() {
-			return containingAsIndexedEmbeddedWithCast;
-		}
-
-		public void setContainingAsIndexedEmbeddedWithCast(Object containingAsIndexedEmbeddedWithCast) {
-			this.containingAsIndexedEmbeddedWithCast = containingAsIndexedEmbeddedWithCast;
-		}
-
-		public String getIndexedField() {
-			return indexedField;
-		}
-
-		public void setIndexedField(String indexedField) {
-			this.indexedField = indexedField;
-		}
-
-		public List<String> getIndexedElementCollectionField() {
-			return indexedElementCollectionField;
-		}
-
-		public void setIndexedElementCollectionField(List<String> indexedElementCollectionField) {
-			this.indexedElementCollectionField = indexedElementCollectionField;
-		}
-
-		public String getNonIndexedField() {
-			return nonIndexedField;
-		}
-
-		public void setNonIndexedField(String nonIndexedField) {
-			this.nonIndexedField = nonIndexedField;
-		}
-
-		public List<String> getNonIndexedElementCollectionField() {
-			return nonIndexedElementCollectionField;
-		}
-
-		public void setNonIndexedElementCollectionField(List<String> nonIndexedElementCollectionField) {
-			this.nonIndexedElementCollectionField = nonIndexedElementCollectionField;
-		}
-
-		public String getFieldUsedInContainedDerivedField1() {
-			return fieldUsedInContainedDerivedField1;
-		}
-
-		public void setFieldUsedInContainedDerivedField1(String fieldUsedInContainedDerivedField1) {
-			this.fieldUsedInContainedDerivedField1 = fieldUsedInContainedDerivedField1;
-		}
-
-		public String getFieldUsedInContainedDerivedField2() {
-			return fieldUsedInContainedDerivedField2;
-		}
-
-		public void setFieldUsedInContainedDerivedField2(String fieldUsedInContainedDerivedField2) {
-			this.fieldUsedInContainedDerivedField2 = fieldUsedInContainedDerivedField2;
-		}
-
-		public String getFieldUsedInCrossEntityDerivedField1() {
-			return fieldUsedInCrossEntityDerivedField1;
-		}
-
-		public void setFieldUsedInCrossEntityDerivedField1(String fieldUsedInCrossEntityDerivedField1) {
-			this.fieldUsedInCrossEntityDerivedField1 = fieldUsedInCrossEntityDerivedField1;
-		}
-
-		public String getFieldUsedInCrossEntityDerivedField2() {
-			return fieldUsedInCrossEntityDerivedField2;
-		}
-
-		public void setFieldUsedInCrossEntityDerivedField2(String fieldUsedInCrossEntityDerivedField2) {
-			this.fieldUsedInCrossEntityDerivedField2 = fieldUsedInCrossEntityDerivedField2;
-		}
-
-		@Transient
-		@GenericField
-		@IndexingDependency(derivedFrom = {
-				@ObjectPath(@PropertyValue(propertyName = "fieldUsedInContainedDerivedField1")),
-				@ObjectPath(@PropertyValue(propertyName = "fieldUsedInContainedDerivedField2"))
-		})
-		public Optional<String> getContainedDerivedField() {
-			return computeDerived( Stream.of( fieldUsedInContainedDerivedField1, fieldUsedInContainedDerivedField2 ) );
-		}
-	}
-
-}
