+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/SingleValuedPropertyAccessor.java
@@ -7,18 +7,33 @@
 package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association;
 
 import java.util.function.BiConsumer;
+import java.util.function.Function;
 
 public final class SingleValuedPropertyAccessor<R, V> implements PropertyAccessor<R, V> {
 
 	private final BiConsumer<R, V> setMethod;
+	private final Function<R, V> getMethod;
 
 	public SingleValuedPropertyAccessor(BiConsumer<R, V> setMethod) {
+		this( setMethod, null );
+	}
+
+	public SingleValuedPropertyAccessor(BiConsumer<R, V> setMethod, Function<R, V> getMethod) {
 		this.setMethod = setMethod;
+		this.getMethod = getMethod;
 	}
 
 	@Override
 	public void set(R root, V value) {
 		setMethod.accept( root, value );
+	}
+
+	@Override
+	public V get(R root) {
+		if ( getMethod == null ) {
+			throw new UnsupportedOperationException();
+		}
+		return getMethod.apply( root );
 	}
 
 	@Override
