+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/single/AutomaticIndexingSingleAssociationIT.java
@@ -1,0 +1,552 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.single;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Stream;
+import javax.persistence.Basic;
+import javax.persistence.CollectionTable;
+import javax.persistence.Column;
+import javax.persistence.ElementCollection;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
+import javax.persistence.Transient;
+
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.AbstractAutomaticIndexingAssociationIT;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.ContainerPrimitives;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.MultiValuedPropertyAccessor;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.PropertyAccessor;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.SingleValuedPropertyAccessor;
+import org.hibernate.search.mapper.pojo.automaticindexing.ReindexOnUpdate;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexingDependency;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.ObjectPath;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.PropertyValue;
+
+/**
+ * Test automatic indexing based on Hibernate ORM entity events
+ * when a ToOne association is involved.
+ * <p>
+ * See {@link AbstractAutomaticIndexingAssociationIT} for more details on how this test is designed.
+ */
+public class AutomaticIndexingSingleAssociationIT extends AbstractAutomaticIndexingSingleAssociationIT<
+		AutomaticIndexingSingleAssociationIT.IndexedEntity,
+		AutomaticIndexingSingleAssociationIT.ContainingEntity,
+		AutomaticIndexingSingleAssociationIT.ContainedEntity
+		> {
+
+	public AutomaticIndexingSingleAssociationIT() {
+		super( new ModelPrimitives() );
+	}
+
+	private static class ModelPrimitives
+			implements SingleValuedAssociationModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity> {
+
+		@Override
+		public String getIndexName() {
+			return IndexedEntity.INDEX;
+		}
+
+		@Override
+		public boolean isMultiValuedAssociation() {
+			return false;
+		}
+
+		@Override
+		public Class<IndexedEntity> getIndexedClass() {
+			return IndexedEntity.class;
+		}
+
+		@Override
+		public Class<ContainingEntity> getContainingClass() {
+			return ContainingEntity.class;
+		}
+
+		@Override
+		public Class<ContainedEntity> getContainedClass() {
+			return ContainedEntity.class;
+		}
+
+		@Override
+		public IndexedEntity newIndexed(int id) {
+			IndexedEntity entity = new IndexedEntity();
+			entity.setId( id );
+			return entity;
+		}
+
+		@Override
+		public ContainingEntity newContaining(int id) {
+			ContainingEntity entity = new ContainingEntity();
+			entity.setId( id );
+			return entity;
+		}
+
+		@Override
+		public ContainedEntity newContained(int id) {
+			ContainedEntity entity = new ContainedEntity();
+			entity.setId( id );
+			return entity;
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainingEntity, String> containingEntityNonIndexedField() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setNonIndexedField );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainingEntity, ContainingEntity> child() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setChild );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainingEntity, ContainingEntity> parent() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setParent );
+		}
+
+		@Override
+		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbedded() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbedded,
+					ContainingEntity::getContainedIndexedEmbedded );
+		}
+
+		@Override
+		public PropertyAccessor<ContainedEntity, ContainingEntity> containingAsIndexedEmbedded() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setContainingAsIndexedEmbedded );
+		}
+
+		@Override
+		public PropertyAccessor<ContainingEntity, ContainedEntity> containedNonIndexedEmbedded() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedNonIndexedEmbedded,
+					ContainingEntity::getContainedNonIndexedEmbedded );
+		}
+
+		@Override
+		public PropertyAccessor<ContainedEntity, ContainingEntity> containingAsNonIndexedEmbedded() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setContainingAsNonIndexedEmbedded );
+		}
+
+		@Override
+		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbeddedShallowReindexOnUpdate );
+		}
+
+		@Override
+		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbeddedNoReindexOnUpdate );
+		}
+
+		@Override
+		public PropertyAccessor<ContainingEntity, ContainedEntity> containedUsedInCrossEntityDerivedProperty() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedUsedInCrossEntityDerivedProperty,
+					ContainingEntity::getContainedUsedInCrossEntityDerivedProperty );
+		}
+
+		@Override
+		public PropertyAccessor<ContainedEntity, ContainingEntity> containingAsUsedInCrossEntityDerivedProperty() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setContainingAsUsedInCrossEntityDerivedProperty );
+		}
+
+		@Override
+		public PropertyAccessor<ContainingEntity, ContainedEntity> containedIndexedEmbeddedWithCast() {
+			return new SingleValuedPropertyAccessor<>( ContainingEntity::setContainedIndexedEmbeddedWithCast );
+		}
+
+		@Override
+		public PropertyAccessor<ContainedEntity, ContainingEntity> containingAsIndexedEmbeddedWithCast() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setContainingAsIndexedEmbeddedWithCast );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainedEntity, String> indexedField() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setIndexedField );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainedEntity, String> nonIndexedField() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setNonIndexedField );
+		}
+
+		@Override
+		public MultiValuedPropertyAccessor<ContainedEntity, String, List<String>> indexedElementCollectionField() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
+					ContainedEntity::getIndexedElementCollectionField,
+					ContainedEntity::setIndexedElementCollectionField );
+		}
+
+		@Override
+		public MultiValuedPropertyAccessor<ContainedEntity, String, List<String>> nonIndexedElementCollectionField() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
+					ContainedEntity::getNonIndexedElementCollectionField,
+					ContainedEntity::setNonIndexedElementCollectionField );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInContainedDerivedField1() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInContainedDerivedField1 );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInContainedDerivedField2() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInContainedDerivedField2 );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInCrossEntityDerivedField1() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInCrossEntityDerivedField1 );
+		}
+
+		@Override
+		public SingleValuedPropertyAccessor<ContainedEntity, String> fieldUsedInCrossEntityDerivedField2() {
+			return new SingleValuedPropertyAccessor<>( ContainedEntity::setFieldUsedInCrossEntityDerivedField2 );
+		}
+	}
+
+	@Entity(name = "containing")
+	public static class ContainingEntity {
+
+		@Id
+		private Integer id;
+
+		private String nonIndexedField;
+
+		@OneToOne
+		private ContainingEntity parent;
+
+		@OneToOne(mappedBy = "parent")
+		@IndexedEmbedded(includePaths = {
+				"containedIndexedEmbedded.indexedField",
+				"containedIndexedEmbedded.indexedElementCollectionField",
+				"containedIndexedEmbedded.containedDerivedField",
+				"containedIndexedEmbeddedShallowReindexOnUpdate.indexedField",
+				"containedIndexedEmbeddedShallowReindexOnUpdate.indexedElementCollectionField",
+				"containedIndexedEmbeddedShallowReindexOnUpdate.containedDerivedField",
+				"containedIndexedEmbeddedNoReindexOnUpdate.indexedField",
+				"containedIndexedEmbeddedNoReindexOnUpdate.indexedElementCollectionField",
+				"containedIndexedEmbeddedNoReindexOnUpdate.containedDerivedField",
+				"containedIndexedEmbeddedWithCast.indexedField",
+				"crossEntityDerivedField"
+		})
+		private ContainingEntity child;
+
+		@OneToOne
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		private ContainedEntity containedIndexedEmbedded;
+
+		@OneToOne
+		private ContainedEntity containedNonIndexedEmbedded;
+
+		@OneToOne
+		@JoinColumn(name = "CIndexedEmbeddedSROU")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
+		private ContainedEntity containedIndexedEmbeddedShallowReindexOnUpdate;
+
+		@OneToOne
+		@JoinColumn(name = "CIndexedEmbeddedNROU")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.NO)
+		private ContainedEntity containedIndexedEmbeddedNoReindexOnUpdate;
+
+		@OneToOne
+		@JoinColumn(name = "CCrossEntityDerived")
+		private ContainedEntity containedUsedInCrossEntityDerivedProperty;
+
+		@OneToOne(targetEntity = ContainedEntity.class)
+		@JoinColumn(name = "CIndexedEmbeddedCast")
+		@IndexedEmbedded(includePaths = { "indexedField" }, targetType = ContainedEntity.class)
+		private Object containedIndexedEmbeddedWithCast;
+
+		public Integer getId() {
+			return id;
+		}
+
+		public void setId(Integer id) {
+			this.id = id;
+		}
+
+		public String getNonIndexedField() {
+			return nonIndexedField;
+		}
+
+		public void setNonIndexedField(String nonIndexedField) {
+			this.nonIndexedField = nonIndexedField;
+		}
+
+		public ContainingEntity getParent() {
+			return parent;
+		}
+
+		public void setParent(ContainingEntity parent) {
+			this.parent = parent;
+		}
+
+		public ContainingEntity getChild() {
+			return child;
+		}
+
+		public void setChild(ContainingEntity child) {
+			this.child = child;
+		}
+
+		public ContainedEntity getContainedIndexedEmbedded() {
+			return containedIndexedEmbedded;
+		}
+
+		public void setContainedIndexedEmbedded(ContainedEntity containedIndexedEmbedded) {
+			this.containedIndexedEmbedded = containedIndexedEmbedded;
+		}
+
+		public ContainedEntity getContainedNonIndexedEmbedded() {
+			return containedNonIndexedEmbedded;
+		}
+
+		public void setContainedNonIndexedEmbedded(ContainedEntity containedNonIndexedEmbedded) {
+			this.containedNonIndexedEmbedded = containedNonIndexedEmbedded;
+		}
+
+		public ContainedEntity getContainedIndexedEmbeddedShallowReindexOnUpdate() {
+			return containedIndexedEmbeddedShallowReindexOnUpdate;
+		}
+
+		public void setContainedIndexedEmbeddedShallowReindexOnUpdate(
+				ContainedEntity containedIndexedEmbeddedShallowReindexOnUpdate) {
+			this.containedIndexedEmbeddedShallowReindexOnUpdate = containedIndexedEmbeddedShallowReindexOnUpdate;
+		}
+
+		public ContainedEntity getContainedIndexedEmbeddedNoReindexOnUpdate() {
+			return containedIndexedEmbeddedNoReindexOnUpdate;
+		}
+
+		public void setContainedIndexedEmbeddedNoReindexOnUpdate(
+				ContainedEntity containedIndexedEmbeddedNoReindexOnUpdate) {
+			this.containedIndexedEmbeddedNoReindexOnUpdate = containedIndexedEmbeddedNoReindexOnUpdate;
+		}
+
+		public ContainedEntity getContainedUsedInCrossEntityDerivedProperty() {
+			return containedUsedInCrossEntityDerivedProperty;
+		}
+
+		public void setContainedUsedInCrossEntityDerivedProperty(
+				ContainedEntity containedUsedInCrossEntityDerivedProperty) {
+			this.containedUsedInCrossEntityDerivedProperty = containedUsedInCrossEntityDerivedProperty;
+		}
+
+		public Object getContainedIndexedEmbeddedWithCast() {
+			return containedIndexedEmbeddedWithCast;
+		}
+
+		public void setContainedIndexedEmbeddedWithCast(Object containedIndexedEmbeddedWithCast) {
+			this.containedIndexedEmbeddedWithCast = containedIndexedEmbeddedWithCast;
+		}
+
+		@Transient
+		@GenericField
+		@IndexingDependency(derivedFrom = {
+				@ObjectPath({
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
+						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField1")
+				}),
+				@ObjectPath({
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
+						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField2")
+				})
+		})
+		public Optional<String> getCrossEntityDerivedField() {
+			return containedUsedInCrossEntityDerivedProperty == null
+					? Optional.empty()
+					: computeDerived( Stream.of(
+							containedUsedInCrossEntityDerivedProperty.getFieldUsedInCrossEntityDerivedField1(),
+							containedUsedInCrossEntityDerivedProperty.getFieldUsedInCrossEntityDerivedField2()
+					) );
+		}
+	}
+
+	@Entity(name = "indexed")
+	@Indexed(index = IndexedEntity.INDEX)
+	public static class IndexedEntity extends ContainingEntity {
+
+		static final String INDEX = "IndexedEntity";
+
+	}
+
+	@Entity(name = "contained")
+	public static class ContainedEntity {
+		@Id
+		private Integer id;
+
+		@OneToOne(mappedBy = "containedIndexedEmbedded")
+		private ContainingEntity containingAsIndexedEmbedded;
+
+		@OneToOne(mappedBy = "containedNonIndexedEmbedded")
+		private ContainingEntity containingAsNonIndexedEmbedded;
+
+		@OneToOne(mappedBy = "containedUsedInCrossEntityDerivedProperty")
+		private ContainingEntity containingAsUsedInCrossEntityDerivedProperty;
+
+		@OneToOne(mappedBy = "containedIndexedEmbeddedWithCast", targetEntity = ContainingEntity.class)
+		private Object containingAsIndexedEmbeddedWithCast;
+
+		@Basic
+		@GenericField
+		private String indexedField;
+
+		@ElementCollection
+		@CollectionTable(name = "indexedECF")
+		@GenericField
+		private List<String> indexedElementCollectionField = new ArrayList<>();
+
+		@Basic
+		@GenericField
+		// Keep this annotation, it should be ignored because the field is not included in the @IndexedEmbedded
+		private String nonIndexedField;
+
+		@ElementCollection
+		@CollectionTable(name = "nonIndexedECF")
+		@Column(name = "nonIndexedECF")
+		@GenericField
+		// Keep this annotation, it should be ignored because the field is not included in the @IndexedEmbedded
+		private List<String> nonIndexedElementCollectionField = new ArrayList<>();
+
+		@Basic // Do not annotate with @GenericField, this would make the test pointless
+		@Column(name = "FUIContainedDF1")
+		private String fieldUsedInContainedDerivedField1;
+
+		@Basic // Do not annotate with @GenericField, this would make the test pointless
+		@Column(name = "FUIContainedDF2")
+		private String fieldUsedInContainedDerivedField2;
+
+		@Basic // Do not annotate with @GenericField, this would make the test pointless
+		@Column(name = "FUICrossEntityDF1")
+		private String fieldUsedInCrossEntityDerivedField1;
+
+		@Basic // Do not annotate with @GenericField, this would make the test pointless
+		@Column(name = "FUICrossEntityDF2")
+		private String fieldUsedInCrossEntityDerivedField2;
+
+		public Integer getId() {
+			return id;
+		}
+
+		public void setId(Integer id) {
+			this.id = id;
+		}
+
+		public ContainingEntity getContainingAsIndexedEmbedded() {
+			return containingAsIndexedEmbedded;
+		}
+
+		public void setContainingAsIndexedEmbedded(ContainingEntity containingAsIndexedEmbedded) {
+			this.containingAsIndexedEmbedded = containingAsIndexedEmbedded;
+		}
+
+		public ContainingEntity getContainingAsNonIndexedEmbedded() {
+			return containingAsNonIndexedEmbedded;
+		}
+
+		public void setContainingAsNonIndexedEmbedded(ContainingEntity containingAsNonIndexedEmbedded) {
+			this.containingAsNonIndexedEmbedded = containingAsNonIndexedEmbedded;
+		}
+
+		public ContainingEntity getContainingAsUsedInCrossEntityDerivedProperty() {
+			return containingAsUsedInCrossEntityDerivedProperty;
+		}
+
+		public void setContainingAsUsedInCrossEntityDerivedProperty(
+				ContainingEntity containingAsUsedInCrossEntityDerivedProperty) {
+			this.containingAsUsedInCrossEntityDerivedProperty = containingAsUsedInCrossEntityDerivedProperty;
+		}
+
+		public Object getContainingAsIndexedEmbeddedWithCast() {
+			return containingAsIndexedEmbeddedWithCast;
+		}
+
+		public void setContainingAsIndexedEmbeddedWithCast(Object containingAsIndexedEmbeddedWithCast) {
+			this.containingAsIndexedEmbeddedWithCast = containingAsIndexedEmbeddedWithCast;
+		}
+
+		public String getIndexedField() {
+			return indexedField;
+		}
+
+		public void setIndexedField(String indexedField) {
+			this.indexedField = indexedField;
+		}
+
+		public List<String> getIndexedElementCollectionField() {
+			return indexedElementCollectionField;
+		}
+
+		public void setIndexedElementCollectionField(List<String> indexedElementCollectionField) {
+			this.indexedElementCollectionField = indexedElementCollectionField;
+		}
+
+		public String getNonIndexedField() {
+			return nonIndexedField;
+		}
+
+		public void setNonIndexedField(String nonIndexedField) {
+			this.nonIndexedField = nonIndexedField;
+		}
+
+		public List<String> getNonIndexedElementCollectionField() {
+			return nonIndexedElementCollectionField;
+		}
+
+		public void setNonIndexedElementCollectionField(List<String> nonIndexedElementCollectionField) {
+			this.nonIndexedElementCollectionField = nonIndexedElementCollectionField;
+		}
+
+		public String getFieldUsedInContainedDerivedField1() {
+			return fieldUsedInContainedDerivedField1;
+		}
+
+		public void setFieldUsedInContainedDerivedField1(String fieldUsedInContainedDerivedField1) {
+			this.fieldUsedInContainedDerivedField1 = fieldUsedInContainedDerivedField1;
+		}
+
+		public String getFieldUsedInContainedDerivedField2() {
+			return fieldUsedInContainedDerivedField2;
+		}
+
+		public void setFieldUsedInContainedDerivedField2(String fieldUsedInContainedDerivedField2) {
+			this.fieldUsedInContainedDerivedField2 = fieldUsedInContainedDerivedField2;
+		}
+
+		public String getFieldUsedInCrossEntityDerivedField1() {
+			return fieldUsedInCrossEntityDerivedField1;
+		}
+
+		public void setFieldUsedInCrossEntityDerivedField1(String fieldUsedInCrossEntityDerivedField1) {
+			this.fieldUsedInCrossEntityDerivedField1 = fieldUsedInCrossEntityDerivedField1;
+		}
+
+		public String getFieldUsedInCrossEntityDerivedField2() {
+			return fieldUsedInCrossEntityDerivedField2;
+		}
+
+		public void setFieldUsedInCrossEntityDerivedField2(String fieldUsedInCrossEntityDerivedField2) {
+			this.fieldUsedInCrossEntityDerivedField2 = fieldUsedInCrossEntityDerivedField2;
+		}
+
+		@Transient
+		@GenericField
+		@IndexingDependency(derivedFrom = {
+				@ObjectPath(@PropertyValue(propertyName = "fieldUsedInContainedDerivedField1")),
+				@ObjectPath(@PropertyValue(propertyName = "fieldUsedInContainedDerivedField2"))
+		})
+		public Optional<String> getContainedDerivedField() {
+			return computeDerived( Stream.of( fieldUsedInContainedDerivedField1, fieldUsedInContainedDerivedField2 ) );
+		}
+	}
+
+}
